CCC-1710 Possibly CCC-1704, CCC-1686 Fix Wrong Socket Handler
There was two event callbacks handlers which was init using On
method. This method can be called via Server (Namespace) or Socket.
Both need their individual backend object to store. Previous
commit made big mistake by removing the Socket's handler and point
it back to the Server's handler. Both handler need to be
seperated. The mistake caused the problem where event callbacks
init with Socket.On method only execute on the last websocket that
triggered the callback event that was registered with Server.On
method and provide the Socket object for the former to register
the event callbacks. For example:

Server.On('connection', func(s Socket) {
	s.On('register', func() {
		// some codes
		s.Emit(...)
	})
})

s.Emit may only be triggerred on websocket that last triggered the
'connection' event as the variable s only got captured by that
last callback and stored under Server.namespace.baseHandler. To
capture every s for every websocket, the callback for 'connection'
need to be stored and use it for that specific websocket. This
commit put back original author code that make this happen.

Also removed the sync.Mutex init for baseHandler as it is
redundant.